docs: refactor and cleanup
