+++ content/assets/stylesheets/home.css
@@ -1,61 +1,36 @@
-/* set background image for the main page */
-[data-md-color-scheme=default] .md-container {
-    background-image: url("/assets/images/moving-firefly-2.svg");
-    background-repeat: no-repeat;
-    background-size: cover;
-    background-position: center;
-}
-
-[data-md-color-scheme=slate] .md-container {
-    background-image: url("/assets/images/moving-firefly-2.svg");
-    background-repeat: no-repeat;
-    background-size: cover;
-    background-position: center;
-}
-
-/* hide markdown content section */
-.md-content {
-    display: none;
-}
-
-/* prevent unused main section from taking up space */
-.md-main {
-    max-height: 0 !important;
-}
-
-/* ensure our primary div takes up the full screen, and align its contents in the center */
+/* ensure our primary div takes up the full screen */
 .mdx-container {
     display: flex;
     flex-grow: 1;
-    flex-direction: column;
+}
+
+/* set background image for the main page */
+.mdx-container {
+    background-image: url("/assets/images/moving-firefly-2.svg");
+    background-repeat: no-repeat;
+    background-size: cover;
+    background-position: center;
 }
 
 /* styling for hero content */
 .mdx-hero {
     display: flex;
     flex-grow: 1;
-    align-items: center;
-    padding-top: 1.5rem;
-    padding-bottom: 1.5rem;
-    padding-left: 1.25rem;
-    padding-right: 1.25rem;
 }
 
 .mdx-hero .mdx-hero__content {
-    margin-left: auto;
-    margin-right: auto;
+    margin: auto;
     text-align: center;
 }
 
 .mdx-hero .mdx-hero__content_inner {
+    padding: 2.0rem 0.75rem;
     max-width: 25rem;
-    margin-left: auto;
-    margin-right: auto;
 }
 
 .mdx-hero .mdx-hero__tag-line {
     margin-bottom: 0.0rem;
-    padding-bottom: 2.0rem;
+    padding-bottom: 1.0rem;
     color: var(--md-typeset-color);
     font-weight: 700;
     font-size: 2.4rem;
@@ -76,7 +51,7 @@
 }
 
 .mdx-hero .mdx-hero__tag-description {
-    padding-bottom: 1.0rem;
+    padding-bottom: 1.4rem;
 }
 
 .mdx-hero .mdx-hero__crafted_by_text {
@@ -84,60 +59,36 @@
     font-size: 70%
 }
 
-.mdx-hero__video {
-    max-width: 38rem;
-    margin-left: auto;
-    margin-right: auto;
+.mdx-hero .mdx-hero__actions {
+    padding-bottom: 1.0rem;
 }
 
-/* styles for action buttons */
-.mdx-hero .mdx-actions {
-    padding-bottom: 2.5rem;
-}
-
-.mdx-hero .md-button {
+.mdx-hero .mdx-hero__actions .md-button {
     margin: .3rem .25rem;
 }
 
-.mdx-hero .md-button .mdx-button-icon {
+.mdx-hero .mdx-hero__actions .md-button .mdx-button-icon {
     padding-left: .3rem;
 }
 
-.mdx-hero .md-button--primary {
+.mdx-hero .mdx-hero__actions .md-button--primary {
     background-color: var(--md-accent-fg-color) !important;
     border-color: var(--md-accent-fg-color);
     color: white !important;
 }
 
-.mdx-hero .md-button--secondary {
+.mdx-hero .mdx-hero__actions .md-button--secondary {
     background-color: var(--md-accent-fg-color--transparent) !important;
     border-color: var(--md-primary-fg-color--light);
     color: var(--md-primary-fg-color--light) !important;
 }
 
-[data-md-color-scheme=default] .mdx-hero .md-button:focus, [data-md-color-scheme=default] .mdx-hero .md-button:hover {
+[data-md-color-scheme=default] .mdx-hero .mdx-hero__actions .md-button:focus,
+[data-md-color-scheme=default] .mdx-hero .mdx-hero__actions .md-button:hover {
     border-color: white;
 }
 
-[data-md-color-scheme=slate] .mdx-hero .md-button:focus, [data-md-color-scheme=slate] .mdx-hero .md-button:hover {
+[data-md-color-scheme=slate] .mdx-hero .mdx-hero__actions .md-button:focus,
+[data-md-color-scheme=slate] .mdx-hero .mdx-hero__actions .md-button:hover {
     border-color: white;
-}
-
-/* styling for video */
-.mdx-hero__video {
-    width: auto;
-}
-
-.mdx-hero__video_inner {
-    position: relative;
-    padding-bottom: 56.25%; /* 56.25% == 16:9 */
-    padding-top: 25px;
-}
-
-.mdx-hero__video_inner iframe {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
 }
