+++ content/index.md
@@ -1,4 +1,5 @@
 ---
+title: deployKF - Your Open ML Platform
 template: home.html
 hide:
   - toc
@@ -7,5 +8,5 @@ schema:
   - "@context": https://schema.org
     "@type": Organization
     url: https://www.deploykf.org/
-    logo: https://www.deploykf.org/assets/images/logo_1/logo.svg
+    logo: https://www.deploykf.org/assets/images/logo_1/logo_wide.svg
 ---
