+++ macros/__init__.py
@@ -1,0 +1,1 @@
+from macros.custom_macros import *
