+++ overrides/main.html
@@ -1,16 +1,21 @@
 {% extends "base.html" %}
 
-<!-- Announcement Bar -->
-{% block announce %}
-  <h3>
-    <b>ANNOUNCEMENT:</b>
-    <a href="https://github.com/deployKF/deployKF/releases/tag/v0.1.1">
-      We are pleased to announce that <strong>deployKF v0.1.1</strong> is now available! 🚀
-    </a>
-  </h3>
+<!-- SEO Data -->
+{% block extrahead %}
+{% include "partials_custom/seo_data.html" %}
 {% endblock %}
 
-<!-- SEO Content -->
-{% block extrahead %}
-  {% include "partials_custom/seo_data.html" %}
+<!-- Sync Announcement Bar -->
+<!-- NOTE: We use the "hero" block because it's inside the "content" object so changes with the page -->
+<!-- NOTE: Scripts inside "content" are run after page change in "instant navigation" mode. -->
+{% block hero %}
+<script>syncAnnouncementBar()</script>
+{% endblock %}
+
+<!-- Announcement Bar -->
+{% block announce %}
+<b>ANNOUNCEMENT: </b>
+<a href="https://github.com/deployKF/deployKF/releases/tag/v0.1.1">
+  We are pleased to announce that <strong>deployKF v0.1.1</strong> is now available! 🚀
+</a>
 {% endblock %}
