+++ overrides/partials/header.html
@@ -1,46 +1,37 @@
-<!-- Determine classes -->
+{#-
+  BASED ON:
+   - SRC: https://github.com/squidfunk/mkdocs-material/blob/9.2.3/src/partials/header.html
+   - OUT: https://github.com/squidfunk/mkdocs-material/blob/9.2.3/material/partials/header.html
+
+  CHANGES:
+   - use `"partials_custom/logo_banner.html"` instead of `"partials/logo.html"`
+   - replace `{{ config.site_name }}` with a logo that only appears on small screens
+-#}
 {% set class = "md-header" %}
 {% if "navigation.tabs.sticky" in features %}
   {% set class = class ~ " md-header--shadow md-header--lifted" %}
 {% elif "navigation.tabs" not in features %}
   {% set class = class ~ " md-header--shadow" %}
 {% endif %}
-
-<!-- Header -->
 <header class="{{ class }}" data-md-component="header">
-  <nav
-    class="md-header__inner md-grid"
-    aria-label="{{ lang.t('header') }}"
-  >
-
-    <!-- Link to home -->
-    <a
-      href="{{ config.extra.homepage | d(nav.homepage.url, true) | url }}"
-      title="{{ config.site_name | e }}"
-      class="md-header__button md-logo"
-      aria-label="{{ config.site_name }}"
-      data-md-component="logo"
-    >
+  <nav class="md-header__inner md-grid" aria-label="{{ lang.t('header') }}">
+    <a href="{{ config.extra.homepage | d(nav.homepage.url, true) | url }}" title="{{ config.site_name | e }}" class="md-header__button md-logo" aria-label="{{ config.site_name }}" data-md-component="logo">
       <!-- ================ BEGIN CHANGE ================ -->
       {% include "partials_custom/logo_banner.html" %}
       <!-- ================= END CHANGE ================= -->
     </a>
-
-    <!-- Button to open drawer -->
     <label class="md-header__button md-icon" for="__drawer">
       {% set icon = config.theme.icon.menu or "material/menu" %}
       {% include ".icons/" ~ icon ~ ".svg" %}
     </label>
-
-    <!-- Header title -->
     <div class="md-header__title" data-md-component="header-title">
       <div class="md-header__ellipsis">
         <div class="md-header__topic">
-          <span class="md-ellipsis">
-            <!-- ================ BEGIN CHANGE ================ -->
-            <!--{{ config.site_name }}-->
-            <!-- ================= END CHANGE ================= -->
-          </span>
+          <!-- ================ BEGIN CHANGE ================ -->
+          <a href="{{ config.extra.homepage | d(nav.homepage.url, true) | url }}" title="{{ config.site_name | e }}" class="md-header__button mdx-logo" aria-label="{{ config.site_name }}" data-md-component="logo">
+            {% include "partials_custom/logo_banner.html" %}
+          </a>
+          <!-- ================= END CHANGE ================= -->
         </div>
         <div class="md-header__topic" data-md-component="header-topic">
           <span class="md-ellipsis">
@@ -53,39 +44,27 @@
         </div>
       </div>
     </div>
-
-    <!-- Color palette toggle -->
     {% if config.theme.palette %}
       {% if not config.theme.palette is mapping %}
         {% include "partials/palette.html" %}
       {% endif %}
     {% endif %}
-
-    <!-- Site language selector -->
     {% if config.extra.alternate %}
       {% include "partials/alternate.html" %}
     {% endif %}
-
-    <!-- Button to open search modal -->
     {% if "material/search" in config.plugins %}
       <label class="md-header__button md-icon" for="__search">
         {% set icon = config.theme.icon.search or "material/magnify" %}
         {% include ".icons/" ~ icon ~ ".svg" %}
       </label>
-
-      <!-- Search interface -->
       {% include "partials/search.html" %}
     {% endif %}
-
-    <!-- Repository information -->
     {% if config.repo_url %}
       <div class="md-header__source">
         {% include "partials/source.html" %}
       </div>
     {% endif %}
   </nav>
-
-  <!-- Navigation tabs (sticky) -->
   {% if "navigation.tabs.sticky" in features %}
     {% if "navigation.tabs" in features %}
       {% include "partials/tabs.html" %}
