+++ overrides/partials/javascripts/announce.html
@@ -1,0 +1,10 @@
+{#-
+  BASED ON:
+   - SRC: https://github.com/squidfunk/mkdocs-material/blob/9.2.3/src/partials/javascripts/announce.html
+   - OUT: https://github.com/squidfunk/mkdocs-material/blob/9.2.3/material/partials/javascripts/announce.html
+
+  CHANGES:
+   - wrap the script in a function called `syncAnnouncementBar()` that we call from the "hero" block in "main.html"
+     to ensure the announcement bar remains hidden after page change in "instant navigation" mode (if dismissed)
+-#}
+<script>function syncAnnouncementBar() {var content,el=document.querySelector("[data-md-component=announce]");el&&(content=el.querySelector(".md-typeset"),__md_hash(content.innerHTML)===__md_get("__announce")&&(el.hidden=!0))}</script>
