+++ overrides/partials/nav.html
@@ -1,6 +1,13 @@
-{% import "partials/nav-item.html" as item with context %}
+{#-
+  BASED ON:
+   - SRC: https://github.com/squidfunk/mkdocs-material/blob/9.2.3/src/partials/nav.html
+   - OUT: https://github.com/squidfunk/mkdocs-material/blob/9.2.3/material/partials/nav.html
 
-<!-- Determine classes -->
+  CHANGES:
+   - use `"partials_custom/logo_wide.html"` instead of `"partials/logo.html"`
+   - remove `{{ config.site_name }}` from label
+-#}
+{% import "partials/nav-item.html" as item with context %}
 {% set class = "md-nav md-nav--primary" %}
 {% if "navigation.tabs" in features %}
   {% set class = class ~ " md-nav--lifted" %}
@@ -8,40 +15,22 @@
 {% if "toc.integrate" in features %}
   {% set class = class ~ " md-nav--integrated" %}
 {% endif %}
-
-<!-- Navigation -->
-<nav
-  class="{{ class }}"
-  aria-label="{{ lang.t('nav') }}"
-  data-md-level="0"
->
-
-  <!-- Site title -->
+<nav class="{{ class }}" aria-label="{{ lang.t('nav') }}" data-md-level="0">
   <label class="md-nav__title" for="__drawer">
-    <a
-      href="{{ config.extra.homepage | d(nav.homepage.url, true) | url }}"
-      title="{{ config.site_name | e }}"
-      class="md-nav__button md-logo"
-      aria-label="{{ config.site_name }}"
-      data-md-component="logo"
-    >
+    <a href="{{ config.extra.homepage | d(nav.homepage.url, true) | url }}" title="{{ config.site_name | e }}" class="md-nav__button md-logo" aria-label="{{ config.site_name }}" data-md-component="logo">
       <!-- ================ BEGIN CHANGE ================ -->
-      {% include "partials_custom/logo_banner.html" %}
+      {% include "partials_custom/logo_wide.html" %}
       <!-- ================= END CHANGE ================= -->
     </a>
     <!-- ================ BEGIN CHANGE ================ -->
     <!--{{ config.site_name }}-->
     <!-- ================= END CHANGE ================= -->
   </label>
-
-  <!-- Repository information -->
   {% if config.repo_url %}
     <div class="md-nav__source">
       {% include "partials/source.html" %}
     </div>
   {% endif %}
-
-  <!-- Navigation list -->
   <ul class="md-nav__list" data-md-scrollfix>
     {% for nav_item in nav %}
       {% set path = "__nav_" ~ loop.index %}
