+++ overrides/partials_custom/logo_banner.html
@@ -1,2 +1,4 @@
-<!-- Banner Logo -->
+{#-
+  A custom partial with the "banner" version of the logo.
+-#}
 <img src="{{ config.theme.logo_banner | url }}" alt="{{ config.site_name }} logo" />
