+++ overrides/partials_custom/logo_wide.html
@@ -1,0 +1,4 @@
+{#-
+  A custom partial with the "wide" version of the logo.
+-#}
+<img src="{{ config.theme.logo_wide | url }}" alt="{{ config.site_name }} logo" />
