+++ overrides/partials_custom/seo_data.html
@@ -1,3 +1,7 @@
+{#-
+  A custom partial which generates the SEO data for a page.
+-#}
+
 {#- ================ #}
 {#- BREADCRUMBS: https://developers.google.com/search/docs/appearance/structured-data/breadcrumb #}
 {#- ================ #}
