+++ requirements.txt
@@ -1,3 +1,4 @@
+# when updating `mkdocs-material`, remember to update `overrides/`
 mkdocs-material==9.2.3
 mkdocs-awesome-pages-plugin==2.9.2
 mkdocs-git-revision-date-localized-plugin==1.2.0
