+++ update_changelogs.sh
@@ -26,7 +26,7 @@ python ./dev/generate_changelog.py \
   --output-hide-sections 'navigation' \
   --include-headings 'Significant Changes' 'New Features' 'Improvements' 'Bug Fixes' 'Documentation' 'Miscellaneous' \
   --include-tag-names 'v[0-9]+\.[0-9]+\.[0-9]+(?:-.+)?' \
-  --write-version-file-path ./overrides/variables/latest_deploykf_version.txt
+  --write-version-file-path ./variables/latest_deploykf_version.txt
 
 # generate the changelog for the deployKF/cli repo (excluding pre-releases)
 python ./dev/generate_changelog.py \
@@ -49,5 +49,5 @@ python ./dev/generate_changelog.py \
   --output-hide-sections 'navigation' \
   --include-headings 'Significant Changes' 'New Features' 'Improvements' 'Bug Fixes' 'Documentation' 'Miscellaneous' \
   --include-tag-names 'v[0-9]+\.[0-9]+\.[0-9]+(?:-.+)?' \
-  --write-version-file-path ./overrides/variables/latest_deploykf_cli_version.txt
+  --write-version-file-path ./variables/latest_deploykf_cli_version.txt
