+++ variables/latest_deploykf_cli_version.txt
