+++ variables/latest_deploykf_version.txt
